Split SvelteImportUtil & SvelteModuleUtil
