+++ src/main/java/dev/blachut/svelte/lang/completion/SvelteInsertHandler.kt
@@ -4,7 +4,7 @@ import com.intellij.codeInsight.completion.InsertHandler
 import com.intellij.codeInsight.completion.InsertionContext
 import com.intellij.codeInsight.lookup.LookupElement
 import dev.blachut.svelte.lang.codeInsight.ComponentLookupObject
-import dev.blachut.svelte.lang.codeInsight.SvelteComponentImporter
+import dev.blachut.svelte.lang.codeInsight.SvelteImportUtil
 
 object SvelteInsertHandler : InsertHandler<LookupElement> {
     override fun handleInsert(context: InsertionContext, item: LookupElement) {
@@ -12,6 +12,6 @@ object SvelteInsertHandler : InsertHandler<LookupElement> {
 
         val componentName = item.lookupString
 
-        SvelteComponentImporter.insertComponentImport(context.editor, context.file, lookupObject.file, componentName, lookupObject.moduleInfo)
+        SvelteImportUtil.insertComponentImport(context.editor, context.file, lookupObject.file, componentName, lookupObject.moduleInfo)
     }
 }
