+++ src/main/java/dev/blachut/svelte/lang/inspections/SvelteImportComponentFix.kt
@@ -9,14 +9,14 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.util.PsiEditorUtil
 import com.intellij.psi.xml.XmlTag
-import dev.blachut.svelte.lang.codeInsight.SvelteComponentImporter
+import dev.blachut.svelte.lang.codeInsight.SvelteImportUtil
 
 // TODO Reuse ImportES6ModuleFix
 class SvelteImportComponentFix(tag: XmlTag, private val quote: String, private val componentName: String, private val info: JSModuleNameInfo,
                                private val currentFile: VirtualFile, private val componentFile: VirtualFile) : LocalQuickFixOnPsiElement(tag) {
     override fun getText(): String {
         val wholeImportWrapQuote = ES6ImportPsiUtil.invertQuote(quote)
-        val importText = SvelteComponentImporter.getImportText(currentFile, componentFile, componentName, quote, info)
+        val importText = SvelteImportUtil.getImportText(currentFile, componentFile, componentName, quote, info)
 
         return "Insert $wholeImportWrapQuote$importText$wholeImportWrapQuote"
     }
@@ -27,6 +27,6 @@ class SvelteImportComponentFix(tag: XmlTag, private val quote: String, private v
 
     override fun invoke(project: Project, file: PsiFile, startElement: PsiElement, endElement: PsiElement) {
         val editor = PsiEditorUtil.Service.getInstance().findEditorByPsiElement(startElement) ?: return
-        SvelteComponentImporter.insertComponentImport(editor, file, componentFile, componentName, info)
+        SvelteImportUtil.insertComponentImport(editor, file, componentFile, componentName, info)
     }
 }
